Store a list of modules installed using modman
After applying this commit, modules will be added and removed from a
newly-created .repos file inside the .modman directory. This file is
intended to be committed to the super-repository, and is roughly
equivalent to the lock file in Composer.

This is only really relevant if you're working on a site from more than
one location, which is often the case when you're working in a
multi-user environment.